requirements.txt
# --- Core Framework ---
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10

# --- AI & Embeddings ---
langchain-google-genai==0.0.5
langchain-huggingface==0.0.1
sentence-transformers==2.2.2

# --- Vector Database ---
chromadb==0.4.22
langchain-chroma==0.1.0

# --- Utilities & Frontend ---
streamlit==1.30.0
python-dotenv==1.0.0
pypdf==4.0.0